Improved api

+++ backend/timemanager/lib/timemanager_web/controllers/user_controller.ex
@@ -18,7 +18,9 @@ defmodule TimemanagerWeb.UserController do
 
   def index(conn, _params) do
     if Timemanager.IsAdmin.is_admin(conn) !== true do
-      render(conn, "error.json", %{error: "You are not allowed to do this action!"})
+      conn
+      |> put_status(:forbidden)
+      |> render("error.json", %{error: "You are not authorized to access this resource"})
     else
       users = Employees.list_users()
       render(conn, "index.json", users: users)
@@ -32,32 +34,56 @@ defmodule TimemanagerWeb.UserController do
 
   def show(conn, %{"userID" => userID}) do
     if Timemanager.IsAdmin.is_admin(conn) !== true do
-      render(conn, "error.json", %{error: "You are not allowed to do this action!"})
+      conn
+      |> put_status(:forbidden)
+      |> render("error.json", %{error: "You are not authorized to access this resource"})
     else
       user = Employees.get_user!(userID)
-      render(conn, "show.json", user: user)
+      if !user do
+        conn
+        |> put_status(:not_found)
+        |> render("404.json")
+      else
+        render(conn, "show.json", user: user)
+      end
     end
   end
 
   def update(conn, %{"userID" => userID, "user" => user_params}) do
     {parsedUserID, ""} = Integer.parse(userID)
     if Timemanager.IsAdmin.is_admin(conn) !== true and Guardian.Plug.current_resource(conn).id !== parsedUserID do
-      render(conn, "error.json", %{error: "You are not allowed to do this action!"})
+      conn
+      |> put_status(:forbidden)
+      |> render("error.json", %{error: "You are not authorized to access this resource"})
     else
       user = Employees.get_user!(userID)
-      with {:ok, %User{} = user} <- Employees.update_user(user, user_params) do
-        render(conn, "show.json", user: user)
+      if !user do
+        conn
+        |> put_status(:not_found)
+        |> render("404.json")
+      else
+        with {:ok, %User{} = user} <- Employees.update_user(user, user_params) do
+          render(conn, "show.json", user: user)
+        end
       end
     end
   end
 
   def delete(conn, %{"userID" => userID}) do
     if Timemanager.IsAdmin.is_admin(conn) !== true do
-      render(conn, "error.json", %{error: "You are not allowed to do this action!"})
+      conn
+      |> put_status(:forbidden)
+      |> render("error.json", %{error: "You are not authorized to access this resource"})
     else
       user = Employees.get_user!(userID)
-      with {:ok, %User{}} <- Employees.delete_user(user) do
-        send_resp(conn, :no_content, "")
+      if !user do
+        conn
+        |> put_status(:not_found)
+        |> render("404.json")
+      else
+        with {:ok, %User{}} <- Employees.delete_user(user) do
+          send_resp(conn, :no_content, "")
+        end
       end
     end
   end
